Remove DEPCACHEAGE and the alias for running go without the garbage collector
The `DEPCACHEAGE` messes up dep when updating recent dependencies so I
had to remove it because for whatever reason the `-update` flag doesn't
automatically update that specific dependency. The cache is either all
on or all off even for simple updates.

We have also mostly moved to go modules so it's not really necessary
anymore.

Also removing the `GOGC=off` trick for speeding up the compiler. I may
add this one back in, but I have no idea if it actually improves
anything anymore.